+++ src/main/java/net/goodnightkimba/wgpg/region/Polygon2D.java
@@ -6,9 +6,9 @@ import java.util.List;
 import com.sk89q.worldedit.BlockVector2D;
 
 public class Polygon2D {
-	private int radius;
+	private double radius;
 	private int vertices = 360;
-	private int offSet = 0;
+	private double offSet = 0;
 	private double centerX = 0;
 	private double centerY = 0;
 
@@ -21,7 +21,7 @@ public class Polygon2D {
 	 * @param centerX Center X coordinate.
 	 * @param centerY Center Y coordinate.
 	 */
-	public Polygon2D(int radius, int vertices, int offset, double centerX, double centerY) {
+	public Polygon2D(double radius, int vertices, double offset, double centerX, double centerY) {
 		this.radius = radius;
 		this.vertices = vertices;
 		this.offSet = offset;
@@ -63,11 +63,11 @@ public class Polygon2D {
 		this.centerY = centerY;
 	}
 
-	public int getOffset() {
+	public double getOffset() {
 		return offSet;
 	}
 
-	public void setOffset(int offset) {
+	public void setOffset(double offset) {
 		this.offSet = offset;
 	}
 
@@ -79,11 +79,11 @@ public class Polygon2D {
 		this.vertices = vertices;
 	}
 
-	public int getRadius() {
+	public double getRadius() {
 		return radius;
 	}
 
-	public void setRadius(int radius) {
+	public void setRadius(double radius) {
 		this.radius = radius;
 	}
 }
